Refactored spaghetti.
Created two new classes BeefmoteServer and UiHandler to handle the
server messaging and the updates to the UI, respectively. Now the
MainActivity class is actually readable!

TODO: I have to find a nicer way of doing network operations in
BeefmoteServer without having to wrap every method in an anonymous
Thread object.