Made changes to make easier progress bar implementation.
Now BeefmoteServer's getTracklist method is not blocking, but instead
relies completely on messaging to signal it's start/end. I also
made some *very* hacky changes in PlaylistUiHandler and
PlaylistActivity in order to handle BeefmoteServer's getTracklist new
behavior. The bottom-line is really this: BeefmoteServer should be
completely reimplemented with an Actor pattern, and handle ALL of its
functionality through a message queue.