Fixed current track highlighting bugs.
Now the highlighting works as intended. We had a bug that made the
highlighting not work correctly when the user made a search (steps to
reproduce: do a search, click a track to play it; now delete the search
and go back to the full tracklist: you'll notice that the track
highlighted is not the right one, because the *position* highlighted
remains the same despite the current track *not* being in the same
position anymore). This was due to not updating the current track
adapter position when the user performs a search.
Also fixed another bug that made the highlighting not work when the
current track was not visible in the RecyclerView. This was due to
not calling the PlaylistRecyclerViewAdapter's notifyItemChanged.