Implemented "Add to playback queue" context menu.
It works but I still have to implement a way of showing the playback
queue to the user. I haven't decided yet between simply adding, e.g.,
the playback queue number before the track's title, or adding another
RecyclerView to show the entire playback queue (this would be the more
powerful option but it might be too clumsy, I don't know).